Production schema change: split anchor_status into is_overlapping and untrimmed_anchor_align_id
is_overlapping is 0 by default for all anchor_aligns, and set to 1 by
RemoveAnchorOverlaps.
untrimmed_anchor_align_id is used to track the untrimmed anchor that is the
origin of a (trimmed) anchor. We can find the untrimmed anchors with IS
NULL and the trimmed anchors with IS NOT NULL.